fix: remove format restriction on ID-Cert serial numbers
Per RFC 5280, section 4.1.2.2[1]:

> The serial number MUST be a positive integer assigned by the CA to
> each certificate.  It MUST be unique for each certificate issued by a
> given CA (i.e., the issuer name and serial number identify a unique
> certificate).  CAs MUST force the serialNumber to be a non-negative
> integer.
>
> Given the uniqueness requirements above, serial numbers can be
> expected to contain long integers.  Certificate users MUST be able to
> handle serialNumber values up to 20 octets.  Conforming CAs MUST NOT
> use serialNumber values longer than 20 octets.

The requirement that serial numbers be 64-bit unsigned integers violates the specification, as serial numbers MUST be
positive, but unsigned numbers may be negative when interpreted as signed integers. It is also unnecessarily limiting,
since the spec allows for serial numbers to be up to 20 bytes large (ie. 160 bits), more than double the size Polyproto
currently permits. Removing this limitation would allow implementers to use identfiers such as UUIDs (prefixed with a
zero-byte to ensure the number remains positive), which would exceed the current 64 bit limit.

No extra requirements have been added, as it is assumed Polyproto's requirements are in addition to those in RFC 5280.

[1]: https://datatracker.ietf.org/doc/html/rfc5280#section-4.1.2.2